gome

+++ solver/src/Gulwani4.hs
@@ -141,7 +141,7 @@ testMain = do
       [( 0 , (1.341, 1))
       ,( 3 , (1.341, 6)) ]
 
-satLambda :: Int -> [String] -> Map.Map BitVector (Float, BitVector) -> IO (Maybe String)
+satLambda :: Int -> [String] -> Map.Map BitVector (Double, BitVector) -> IO (Maybe String)
 satLambda probSize opStrs  exampleMap = do
   let 
     examples :: [(SBitVector, SBitVector)]
